+++ src/data/imbues/strange-and-unusual/index.ts
@@ -1,0 +1,17 @@
+export function createStrangeAndUnusualImbues() {
+    return [
+        // armor
+        // createImbueSpiked(),
+        // shield
+        // createImbueMirrored(),
+        // weapon
+        // ...createImbueAir(),
+        // ...createImbueDarkness(),
+        // ...createImbueEarth(),
+        // createImbueElongating(),
+        // ...createImbueLight(),
+        // createImbueThrowing(),
+        // createImbueWater(),
+        // createImbueWood()
+    ];
+}